contact-form/requirements.txt
uvicorn
fastapi
sendgrid
dotenv
python-dotenv
opentelemetry-distro
opentelemetry-exporter-otlp
opentelemetry-instrumentation-fastapi
opentelemetry-sdk


